@@ -1,30 +1,17 @@
 ﻿using SoapHttpClient.Enums;
 using System.Xml.Linq;
 
-namespace SoapHttpClient.DTO
+namespace SoapHttpClient.DTO;
+
+public record SoapMessageConfiguration(SoapVersion SoapVersion)
 {
-    public class SoapMessageConfiguration
-    {
-        public string MediaType { get; }
+    public string MediaType =>
+        SoapVersion == SoapVersion.Soap11
+            ? "text/xml"
+            : "application/soap+xml";
 
-        public XNamespace Schema { get; }
-
-        public SoapVersion SoapVersion { get; }
-
-        public SoapMessageConfiguration(SoapVersion soapVersion)
-        {
-            if (soapVersion == SoapVersion.Soap11)
-            {
-                MediaType = "text/xml";
-                Schema = "http://schemas.xmlsoap.org/soap/envelope/";
-            }
-            else
-            {
-                MediaType = "application/soap+xml";
-                Schema = "http://www.w3.org/2003/05/soap-envelope";
-            }
-
-            SoapVersion = soapVersion;
-        }
-    }
+    public XNamespace Schema =>
+        SoapVersion == SoapVersion.Soap11
+            ? "http://schemas.xmlsoap.org/soap/envelope/"
+            : "http://www.w3.org/2003/05/soap-envelope";
 }
