@@ -1,8 +1,4 @@
 ﻿using SoapHttpClient.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
 using System.Xml.Linq;
 
 namespace SoapHttpClient
@@ -23,9 +19,9 @@ namespace SoapHttpClient
             Uri endpoint,
             SoapVersion soapVersion,
             object body,
-            object header = null,
-            IXElementSerializer xElementSerializer = null,
-            string action = null)
+            object? header = null,
+            IXElementSerializer? xElementSerializer = null,
+            string? action = null)
         {
             if (xElementSerializer == null)
                 xElementSerializer = new XElementSerializer();
@@ -33,7 +29,7 @@ namespace SoapHttpClient
             return client.Post(endpoint,
                 soapVersion,
                 xElementSerializer.Serialize(body),
-                header != null ? new[] { xElementSerializer.Serialize(header) } : default(IEnumerable<XElement>));
+                header != null ? new[] { xElementSerializer.Serialize(header) } : Enumerable.Empty<XElement>());
         }
 
         /// <summary>
@@ -51,8 +47,8 @@ namespace SoapHttpClient
             SoapVersion soapVersion,
             IEnumerable<object> bodies,
             object header,
-            IXElementSerializer xElementSerializer = null,
-            string action = null)
+            IXElementSerializer? xElementSerializer = null,
+            string? action = null)
         {
             if (xElementSerializer == null)
                 xElementSerializer = new XElementSerializer();
@@ -60,7 +56,7 @@ namespace SoapHttpClient
             return client.Post(endpoint,
                 soapVersion,
                 bodies.Select(e => xElementSerializer.Serialize(e)),
-                header != null ? xElementSerializer.Serialize(header) : default(XElement),
+                xElementSerializer.Serialize(header),
                 action);
         }
 
@@ -79,8 +75,8 @@ namespace SoapHttpClient
             SoapVersion soapVersion,
             object body,
             IEnumerable<object> headers,
-            IXElementSerializer xElementSerializer = null,
-            string action = null)
+            IXElementSerializer? xElementSerializer = null,
+            string? action = null)
         {
             if (xElementSerializer == null)
                 xElementSerializer = new XElementSerializer();
