@@ -1,9 +1,4 @@
 ﻿using SoapHttpClient.Enums;
-using System;
-using System.Collections.Generic;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
 using System.Xml.Linq;
 
 namespace SoapHttpClient
@@ -18,6 +13,6 @@ namespace SoapHttpClient
         /// <param name="bodies">The body of the SOAP message.</param>
         /// <param name="headers">The header of the SOAP message.</param>
         /// <param name="action">The SOAPAction of the SOAP message.</param>
-        Task<HttpResponseMessage> PostAsync(Uri endpoint, SoapVersion soapVersion, IEnumerable<XElement> bodies, IEnumerable<XElement> headers = null, string action = null, CancellationToken cancellationToken = default(CancellationToken));
+        Task<HttpResponseMessage> PostAsync(Uri endpoint, SoapVersion soapVersion, IEnumerable<XElement> bodies, IEnumerable<XElement>? headers = null, string? action = null, CancellationToken cancellationToken = default(CancellationToken));
     }
 }
